64bit issues and compiler warnings
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/java/trunk@36094 72102866-910b-0410-8b05-ffd578937521

+++ Source/JIGSMapper.m
@@ -120,7 +120,7 @@ BOOL _JIGSProxyJavaIsEqual (NSMapTable *table, const void *a, const void *b)
  * A solution would be for JNI to provide a hash function for
  * jobjects.  */
 
-unsigned int _JIGSProxyJavaHash (NSMapTable *table, const void *a)
+NSUInteger _JIGSProxyJavaHash (NSMapTable *table, const void *a)
 {
   return 1;
 }
